% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cut_much}
\alias{cut_much}
\title{cut data}
\usage{
cut_much(
  deg_data,
  x,
  y,
  cut_FC = 1,
  cut_P = 0.05,
  label = c("Down", "Stable", "Up"),
  label_ns = "Stable"
)
}
\arguments{
\item{deg_data}{a DEG data frame contains logFC and p value}

\item{x}{which column is log FC}

\item{y}{which column is P value}

\item{cut_FC}{a single number character or numeric vector in threshold value of log FC}

\item{cut_P}{threshold value of P value, can set for every cut_FC number in numeric vector format}

\item{label}{symbol word for groups}

\item{label_ns}{which group is the stable group}
}
\value{
add group column in data frame
}
\description{
group data for plot DEG volcano by FDR and logFC
}
\examples{
cut_much(DEG_df, x = "log2FoldChange", y = "pvalue", cut_FC = 1,cut_P = 0.05,
label = c("Down","Stable","Up"), label_ns = "Stable")
}
