% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deg_data}
\alias{deg_data}
\title{An example dataset}
\format{
An object of class \code{data.frame} with 24952 rows and 6 columns.
}
\usage{
deg_data
}
\description{
a list ordered by samples in counts_data
}
\keyword{datasets}
