% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_DEseq2.R
\name{deg_DESeq2}
\alias{deg_DESeq2}
\title{Run DEseq2}
\usage{
deg_DESeq2(
  counts_data,
  group_list,
  parallel = F,
  test_group,
  control_group,
  qc = TRUE,
  x,
  y,
  dir = ".",
  prefix = "2-DEG_DEseq2"
)
}
\arguments{
\item{counts_data}{a counts data frame of rows in genes and columns in samples}

\item{group_list}{a list ordered by samples in counts_data}

\item{parallel}{if FALSE, no parallelization. if TRUE, parallel execution using BiocParallel}

\item{test_group}{the name of the numerator level for the fold change (Test group)}

\item{control_group}{the name of the denominator level for the fold change (Control group)}

\item{qc}{qc plots}

\item{x}{which column is log FC}

\item{y}{which column is P value}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}
}
\value{
a directory contains figures and csv files and a deg data frame
}
\description{
A integrated function for run DEseq2 in a counts data and return results files.
}
\examples{
\dontrun{
deg_DESeq2(counts_input,group_list,
          test_group = "T", control_group = "C", qc = TRUE,
           x = "log2FoldChange", y = "pvalue",
           dir = tempdir(), prefix = "2-DEG_DEseq2")
}
}
