% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_edgeR.R
\name{deg_edgeR}
\alias{deg_edgeR}
\title{Run degeR}
\usage{
deg_edgeR(
  counts_data,
  group_list,
  cg,
  x = "logFC",
  y = "PValue",
  dir = ".",
  prefix = "2-DEG_edgeR"
)
}
\arguments{
\item{counts_data}{a counts data frame of rows in genes and columns in samples}

\item{group_list}{a list ordered by samples in counts_data}

\item{cg}{the name of the denominator level for the fold change (Control group)}

\item{x}{which column is log FC}

\item{y}{which column is P value}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}
}
\value{
a directory contains figures and csv files  and a deg data frame
}
\description{
A integrated function for run edgeR in a counts data and return results files.
}
\examples{
deg_edgeR(counts_input,group_list,
          cg = "C",
           x = "logFC", y = "PValue",
           dir = tempdir(), prefix = "2-DEG_edgeR")
}
