% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_limma.R
\name{deg_limma}
\alias{deg_limma}
\title{Run limma}
\usage{
deg_limma(
  counts_data,
  group_list,
  tg,
  cg,
  x,
  y,
  dir = ".",
  prefix = "2-DEG_limma"
)
}
\arguments{
\item{counts_data}{a counts data frame of rows in genes and columns in samples}

\item{group_list}{a list ordered by samples in counts_data}

\item{tg}{the name of the numerator level for the fold change (Test group)}

\item{cg}{the name of the denominator level for the fold change (Control group)}

\item{x}{which column is log FC}

\item{y}{which column is P value}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}
}
\value{
a directory contains figures and csv files  and a deg data frame
}
\description{
A integrated function for run limma in a counts data and return results files.
}
\examples{
deg_limma(counts_input,group_list,
          tg = "T", cg = "C",
           x = "logFC", y = "P.Value",
           dir = tempdir(), prefix = "2-DEG_limma")
}
