% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_run.R
\name{deg_run}
\alias{deg_run}
\title{run DEG}
\usage{
deg_run(
  counts_data,
  group_list,
  test_group = "T",
  control_group = "C",
  dir,
  parallel = FALSE
)
}
\arguments{
\item{counts_data}{a counts data frame of rows in genes and columns in samples}

\item{group_list}{a list ordered by samples in counts_data}

\item{test_group}{the name of the numerator level for the fold change (Test group)}

\item{control_group}{the name of the denominator level for the fold change (Control group)}

\item{dir}{a directory to store results}

\item{parallel}{if FALSE, no parallelization. if TRUE, parallel execution using BiocParallel}
}
\value{
a csv file and a DEG_container ob
}
\description{
run DEG module in limma, DESeq2 and edgeR
}
\examples{
\dontrun{
deg_run(counts_input,group_list,test_group = "T", control_group = "C",dir = tempdir())
}
}
