% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_gseplot.R
\name{enhance_gseplot}
\alias{enhance_gseplot}
\title{plot gseKEGG plot}
\usage{
enhance_gseplot(data, top = 10, pvalue_cut = 0.1, enrichmentScore_cut = 0.5)
}
\arguments{
\item{data}{output from enrich_gseKEGG}

\item{top}{filter top by pvalue}

\item{pvalue_cut}{filter cut of pvalue}

\item{enrichmentScore_cut}{filter cut of enrichmentScore}
}
\value{
a list contains up and down
}
\description{
plot gseKEGG plots by up and down
}
\examples{
\dontrun{
gsekegg_res <- enrich_gsekegg(DEG_df,x = "log2FoldChange")
plots_l <- enhance_gseplot(gsekegg_res)
}
}
