% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_volcano.R
\name{enhance_volcano}
\alias{enhance_volcano}
\title{enhance_volcano}
\usage{
enhance_volcano(
  deg_data,
  x,
  y,
  label = c("Down", "Stable", "Up"),
  label_ns = "Stable",
  palette = c("#66c2a5", "grey50", "#fc8d62"),
  cut_FC = "auto",
  cut_P = 0.05,
  top = 10,
  size = 2,
  expand = c(0.25, 0.25),
  genes_list = "top",
  highlight = NULL
)
}
\arguments{
\item{deg_data}{a DEG data frame contains logFC and p value}

\item{x}{which column is log FC}

\item{y}{which column is P value}

\item{label}{symbol word for groups,default is `c("Down","Stable","Up")`}

\item{label_ns}{which group is the stable group}

\item{palette}{color for every label}

\item{cut_FC}{a numeric vector in threshold value of log FC}

\item{cut_P}{threshold value of P value, can set for every cut_FC number in numeric vector format}

\item{top}{a single number or a length of 2 numeric vector, if 2 numeric vector, first one is top max logFC.}

\item{size}{a single number of font size}

\item{expand}{if labels in plot display wrong, try increase it.}

\item{genes_list}{a gene names character vectors;}

\item{highlight}{a gene names character vectors; default NULL}
}
\value{
a ggplot ob
}
\description{
an enhance_colcano function for DEG volcano plot
}
\examples{
enhance_volcano(DEG_df,x = 'log2FoldChange', y = 'pvalue',
                label = c("Down","Stable","Up"), label_ns = "Stable",
                palette = c('#66c2a5', 'grey50', '#fc8d62'),
                cut_FC = "auto",cut_P = 0.05,top = 10, size = 2.0,expand = c(0.25,0.25),
                genes_list = "top", highlight = NULL)
}
