% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erich_kegg.R
\name{enrich_kegg}
\alias{enrich_kegg}
\title{enrich_go}
\usage{
enrich_kegg(
  deg_data,
  x,
  y,
  cut_FC = 1,
  cut_P = 0.05,
  organism = "hsa",
  keyType = "kegg",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  use_internal_data = FALSE,
  label = c("Down", "Stable", "Up"),
  label_ns = "Stable",
  mc.cores = 1L
)
}
\arguments{
\item{deg_data}{a DEG data frame contains logFC and p value}

\item{x}{which column is log FC}

\item{y}{which column is P value}

\item{cut_FC}{a single number character or numeric vector in threshold value of log FC}

\item{cut_P}{threshold value of P value, can set for every cut_FC number in numeric vector format}

\item{organism}{supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html'}

\item{keyType}{one of "kegg", 'ncbi-geneid', 'ncib-proteinid' and 'uniprot'}

\item{pvalueCutoff}{adjusted pvalue cutoff on enrichment tests to report}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{qvalueCutoff}{qvalue cutoff on enrichment tests to report as significant. Tests must pass i) pvalueCutoff on unadjusted pvalues, ii) pvalueCutoff on adjusted pvalues and iii) qvalueCutoff on qvalues to be reported.}

\item{use_internal_data}{logical, use KEGG.db or latest online KEGG data}

\item{label}{symbol word for groups}

\item{label_ns}{which group is the stable group}

\item{mc.cores}{param for mclapply, choose cores to use in non-Windows machine}
}
\value{
enrichKEGG result
}
\description{
a funtion group deg data frame and make enrichment analysis in one step
}
\examples{
\dontrun{
enrich_kegg(deg_data = DEG_df, x = "log2FoldChange", y = "pvalue")
}
}
