% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_run.R
\name{enrichgesKEGG_run}
\alias{enrichgesKEGG_run}
\title{run enrichgesKEGG}
\usage{
enrichgesKEGG_run(
  deg_data,
  x,
  dir = ".",
  prefix = "5-EnrichgseKEGG",
  pvalue_cut = 0.1,
  enrichmentScore_cut = 0.5,
  top = -10,
  organism = "hsa",
  keyType = "kegg",
  OrgDb = "org.Hs.eg.db",
  exponent = 1,
  minGSSize = 10,
  maxGSSize = 500,
  eps = 1e-10,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  verbose = TRUE,
  use_internal_data = FALSE,
  seed = FALSE,
  by = "fgsea"
)
}
\arguments{
\item{deg_data}{a DEG data frame contains logFC and p value}

\item{x}{which column is log FC}

\item{dir}{where to save results files}

\item{prefix}{a prefix of file names in this step}

\item{pvalue_cut}{filter by pvalue}

\item{enrichmentScore_cut}{filter by enrichmentScore}

\item{top}{top rows of up and down}

\item{organism}{supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html'}

\item{keyType}{one of "kegg", 'ncbi-geneid', 'ncib-proteinid' and 'uniprot'}

\item{OrgDb}{OrgDb}

\item{exponent}{weight of each step}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{eps}{This parameter sets the boundary for calculating the p value.}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}

\item{verbose}{print message or not}

\item{use_internal_data}{logical, use KEGG.db or latest online KEGG data}

\item{seed}{logical}

\item{by}{one of 'fgsea' or 'DOSE'}

\item{...}{more param for ggsave}
}
\value{
a list of file
}
\description{
run enrichgesKEGG and make output files
}
\examples{
\dontrun{
enrichgesKEGG_run(deg_data = DEG_df, x = "log2FoldChange", dir = tempdir(),eps = 0)
}
}
