% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_run.R
\name{enrichKEGG_run}
\alias{enrichKEGG_run}
\title{Enrichment analysis of kegg}
\usage{
enrichKEGG_run(
  deg_data,
  x,
  y,
  cut_FC = 1,
  cut_P = 0.05,
  top = 10,
  dir = ".",
  prefix = "4-EnrichKEGG",
  organism = "hsa",
  keyType = "kegg",
  OrgDb = "org.Hs.eg.db",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  use_internal_data = FALSE,
  label = c("Down", "Stable", "Up"),
  label_ns = "Stable",
  mc.cores = 1L,
  down_label = "Down"
)
}
\arguments{
\item{deg_data}{a DEG data frame contains logFC and p value}

\item{x}{which column is log FC}

\item{y}{which column is P value}

\item{cut_FC}{a single number character or numeric vector in threshold value of log FC}

\item{cut_P}{threshold value of P value, can set for every cut_FC number in numeric vector format}

\item{top}{top rows for up and down}

\item{dir}{where to save results files}

\item{prefix}{a prefix of file names in this step}

\item{organism}{supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html'}

\item{keyType}{one of "kegg", 'ncbi-geneid', 'ncib-proteinid' and 'uniprot'}

\item{OrgDb}{OrgDb}

\item{pvalueCutoff}{adjusted pvalue cutoff on enrichment tests to report}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{qvalueCutoff}{qvalue cutoff on enrichment tests to report as significant. Tests must pass i) pvalueCutoff on unadjusted pvalues, ii) pvalueCutoff on adjusted pvalues and iii) qvalueCutoff on qvalues to be reported.}

\item{use_internal_data}{logical, use KEGG.db or latest online KEGG data}

\item{label}{symbol word for groups}

\item{label_ns}{which group is the stable group}

\item{mc.cores}{param for mclapply, choose cores to use in non-Windows machine}

\item{down_label}{which one is Down}
}
\value{
a list result files of KEGG
}
\description{
run enrich_kegg and plot results
}
\examples{
\dontrun{
enrichKEGG_run(deg_data = DEG_df, x = "log2FoldChange", y = "pvalue", cut_FC = 1,
cut_P = 0.05, top = 10)
}
}
