% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_barplot.R
\name{geskegg_barplot}
\alias{geskegg_barplot}
\title{gseKEGG bar plot}
\usage{
geskegg_barplot(data, pvalue_cut = 0.1, enrichmentScore_cut = 0.5, top = -10)
}
\arguments{
\item{data}{a enrich_gsekegg result}

\item{pvalue_cut}{filter by pvalue}

\item{enrichmentScore_cut}{filter by enrichmentScore}

\item{top}{top rows of up and down}
}
\value{
ggplot ob
}
\description{
plot for enrich_gsekegg result
}
\examples{
\dontrun{
test <- enrich_gsekegg(DEG_df,x = "log2FoldChange")
geskegg_barplot(test)
}
}
