% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_barplot.R
\name{kegg_barplot}
\alias{kegg_barplot}
\title{KEGG bar plot}
\usage{
kegg_barplot(data, top = -10, down_label = "Down")
}
\arguments{
\item{data}{a enrich_kegg result, list format contains Up and Down}

\item{top}{top rows of up and down}

\item{down_label}{which one is Down}
}
\value{
a ggplot OB
}
\description{
plot for enrich_kegg result
}
\examples{
\dontrun{
test <- enrich_kegg(deg_data = DEG_df, x = "log2FoldChange", y = "pvalue")
kegg_barplot(test)
}
}
