% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_DESeq2.R
\name{run_DESeq2}
\alias{run_DESeq2}
\title{Basic produce of DESeq2}
\usage{
run_DESeq2(
  counts_data,
  group_list,
  tg,
  cg,
  qc = TRUE,
  dir = ".",
  prefix = "2-DEG_DEseq2"
)
}
\arguments{
\item{counts_data}{a counts data frame of rows in genes and columns in samples}

\item{group_list}{a character vector ordered by samples in counts_data}

\item{tg}{the name of the numerator level for the fold change (Test group)}

\item{cg}{the name of the denominator level for the fold change (Control group)}

\item{qc}{qc plots}
}
\value{
a DEG data frame
}
\description{
A basic function to get data and produce results of DESeq2
}
\examples{
run_DESeq2(counts_input, group_list,tg = "T", cg = "C", dir = tempdir())
}
