% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_edgeR.R
\name{run_edgeR}
\alias{run_edgeR}
\title{Basic produce of edgeR}
\usage{
run_edgeR(counts_data, group_list, cg)
}
\arguments{
\item{counts_data}{a counts data frame of rows in genes and columns in samples}

\item{group_list}{a character vector ordered by samples in counts_data}

\item{cg}{the name of the denominator level for the fold change (Control group)}
}
\value{
a DEG data frame
}
\description{
A basic function to get data and produce results of edgeR
}
\examples{
run_edgeR(counts_input, group_list, cg= "C")
}
