% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_limma.R
\name{run_limma}
\alias{run_limma}
\title{Basic produce of limma}
\usage{
run_limma(counts_data, group_list, cg, tg)
}
\arguments{
\item{counts_data}{a counts data frame of rows in genes and columns in samples}

\item{group_list}{a character vector ordered by samples in counts_data}

\item{cg}{the name of the denominator level for the fold change (Control group)}

\item{tg}{the name of the numerator level for the fold change (Test group)}
}
\value{
a DEG data frame
}
\description{
A basic function to get data and produce results of limma
}
\examples{
run_limma(counts_input, group_list, cg= "C", tg = "T")
}
