% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_volcano.R
\name{top_deg}
\alias{top_deg}
\title{Choose top genes from DEG data frame}
\usage{
top_deg(deg_data, x, y, top, cut_P = 0.05)
}
\arguments{
\item{deg_data}{a DEG data frame contains logFC and p value}

\item{x}{which column is log FC}

\item{y}{which column is P value}

\item{top}{a single number or a length of 2 numeric vector, if 2 numeric vector, first one is top max logFC.}

\item{cut_P}{a single number for threshold value of P value}
}
\value{
a character vector of top genes
}
\description{
For example, if top = 50, head 50 and tail 50 will return from a
ordered DEG data frame after filtering by threshold value of P value
}
\examples{
top_deg(DEG_df, x = "log2FoldChange", y = "pvalue", 50, 0.05)
}
