% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_heatmap.R
\name{enhance_heatmap}
\alias{enhance_heatmap}
\title{Heatmap for DEG data frame}
\usage{
enhance_heatmap(
  counts_data,
  deg_data,
  group_list,
  x,
  y,
  top = 50,
  cut_P = 0.05,
  dir = ".",
  prefix = "2-DEG",
  palette = RColorBrewer::brewer.pal(3, "Set2")[1:2]
)
}
\arguments{
\item{counts_data}{a counts data frame of rows in genes and columns in samples}

\item{deg_data}{a DEG data frame contains logFC and p value}

\item{group_list}{a character vector ordered by samples in counts_data}

\item{x}{which column is log FC}

\item{y}{which column is P value}

\item{top}{a single number or a length of 2 numeric vector, if 2 numeric vector, first one is top max logFC.}

\item{cut_P}{threshold value of P value, can set for every cut_FC number in numeric vector format}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{palette}{a color palette for plots}
}
\value{
a heatmap plot file
}
\description{
default will return a top 100 deg heatmap in p value = 0.05
}
\examples{
enhance_heatmap(counts_input, DEG_df, group_list,
    x = "log2FoldChange", y = "pvalue", dir = tempdir())
}
