% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_DEseq2.R
\name{run_DESeq2}
\alias{run_DESeq2}
\title{Basic produce of DESeq2}
\usage{
run_DESeq2(
  counts_data,
  group_list,
  test_group,
  control_group,
  qc = TRUE,
  dir = ".",
  prefix = "2-DEG_DEseq2",
  parallel = FALSE
)
}
\arguments{
\item{counts_data}{a counts data frame of rows in genes and columns in samples}

\item{group_list}{a character vector ordered by samples in counts_data}

\item{test_group}{the name of the numerator level for the fold change (Test group)}

\item{control_group}{the name of the denominator level for the fold change (Control group)}

\item{qc}{qc plots}

\item{dir}{a directory to store results}

\item{prefix}{a prefix of file names in this step}

\item{parallel}{if FALSE, no parallelization. if TRUE, parallel execution using BiocParallel}
}
\value{
a DEG data frame
}
\description{
A basic function to get data and produce results of DESeq2
}
\examples{
\dontrun{
run_DESeq2(counts_input, group_list,test_group = "T", control_group = "C", dir = tempdir())
}
}
