% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAll.R
\name{runAll}
\alias{runAll}
\title{run all step in default setting}
\usage{
runAll(
  count_data,
  group_list,
  OrgDb = "org.Hs.eg.db",
  dir = ".",
  test_group = "T",
  control_group = "C",
  parallel = FALSE
)
}
\arguments{
\item{count_data}{a counts data frame of rows in genes and columns in samples}

\item{group_list}{a list ordered by samples in counts_data}

\item{OrgDb}{OrgDb}

\item{dir}{a directory to store results}

\item{test_group}{which one is test group in your group_list}

\item{control_group}{which one is control group in your group_list}

\item{parallel}{if FALSE, no parallelization. if TRUE, parallel execution using BiocParallel}
}
\value{
a dir contains all results
}
\description{
Only need count data, group, OrgDb, dir
}
\examples{
\dontrun{
runAll(count_data = counts_input, group_list = group_list, OrgDb = 'org.Hs.eg.db', dir = tempdir())
}
}
